There isn't really a good point to make a commit here since this is just working through the not-implemented errors one by one.
A system of running the parse tree has been implemented (just not for everything yet) and the Hello World part of the document now works.

Exressions and Executables have a .run() method that returns an instance of a Value for Expressions, and an instance of StatusWrapper for Executables. The former are the actual value of the expression as it is resolved. The latter convey directions such as RETURN, BREAK, CONTINUE that can be forwarded. Both can be overloaded to convey error information to build a stack trace.