Add support for SDL2 via the PyGame library.
This will gracefully ignore the pygame import if it is not present.
If any of the game library functionality is used, a message to prompt the user to run `pip install pygame` is presented on the command line.

Currently supports opening a window, checking if the window is closed, filling the screen with a solid color, drawing a solid colored rectangle, and throttling the frame rate.

Eventual goal is to have Tetris by EOD tomorrow.