Create VOTING.md
You can now vote on the team projects you think deserve recognition. To do so, go to the corresponding issue for that team. Each issue starts with "Team: " and then the name of the team.

In that issue, using the upvote (thumbs up) to vote for the submission. Other emoji will be ignored.

Please note, voting is just to highlight projects you think should be highlighted. Please don't use this as a space for critiques of the projects unless the team requests them.